% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probaverse-package.R
\docType{package}
\name{probaverse-package}
\alias{probaverse}
\alias{probaverse-package}
\title{probaverse: Install and Load the 'probaverse' Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'probaverse' is a suite of packages designed to facilitate creating advanced statistical models through probability distributions. These packages work best when loaded together because they share a common design philosophy and focus on different aspects of developing statistical models. Inspired by the 'tidyverse' package, the 'probaverse' package makes it easy to load the entire suite of 'probaverse' packages together.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://probaverse.probaverse.com/}
  \item \url{https://github.com/probaverse/probaverse}
  \item Report bugs at \url{https://github.com/probaverse/probaverse/issues}
}

}
\author{
\strong{Maintainer}: Vincenzo Coia \email{vincenzo.coia@gmail.com} (\href{https://orcid.org/0000-0002-2930-0723}{ORCID}) [copyright holder]

}
\keyword{internal}
